+++ test/mocks/events.js
@@ -1,12 +1,10 @@
 module.exports = function() {
 
-  var self = this;
-
   return {
-    on : self.sandbox.spy(),
-    one : self.sandbox.spy(),
-    trigger : self.sandbox.spy(),
-    unbind : self.sandbox.spy()
+    on : this.sandbox.spy(),
+    one : this.sandbox.spy(),
+    trigger : this.sandbox.spy(),
+    unbind : this.sandbox.spy()
   };
 
 };
